+++ src/msw/window.cpp
@@ -4887,7 +4887,7 @@ static void ScaleCoordIfSet(int& coord, float scaleFactor)
     if ( coord != wxDefaultCoord )
     {
         const float coordScaled = coord * scaleFactor;
-        coord = scaleFactor > 1.0 ? ceil(coordScaled) : floor(coordScaled);
+        coord = int(scaleFactor > 1 ? std::ceil(coordScaled) : std::floor(coordScaled));
     }
 }
