+++ src/xrc/xh_spin.cpp
@@ -67,7 +67,7 @@ bool wxSpinButtonXmlHandler::CanHandle(wxXmlNode *node)
 
 #include "wx/spinctrl.h"
 
-static const float DEFAULT_INC = 1.;
+static const float DEFAULT_INC = 1;
 
 static void AddSpinCtrlStyles(wxXmlResourceHandler& handler)
 {
@@ -134,10 +134,10 @@ wxObject *wxSpinCtrlDoubleXmlHandler::DoCreateResource()
                     GetText(wxS("value")),
                     GetPosition(), GetSize(),
                     GetStyle(wxS("style"), wxSP_ARROW_KEYS),
-                    GetFloat(wxS("min"), (float)DEFAULT_MIN),
-                    GetFloat(wxS("max"), (float)DEFAULT_MAX),
-                    GetFloat(wxS("value"), (float)DEFAULT_VALUE),
-                    GetFloat(wxS("inc"), DEFAULT_INC),
+                    double(GetFloat(wxS("min"), DEFAULT_MIN)),
+                    double(GetFloat(wxS("max"), DEFAULT_MAX)),
+                    double(GetFloat(wxS("value"), DEFAULT_VALUE)),
+                    double(GetFloat(wxS("inc"), DEFAULT_INC)),
                     GetName());
 
     SetupWindow(control);
